Speed up tests by preventing costly `setRepository()` calls
